Rename the `Version` class to `Platform`

+++ scripts/Dist.hx
@@ -1,25 +1,24 @@
 //! --class-path src --library tink_core
-import php_index.base.Platform;
 import sys.FileSystem;
 
 /** Packages the project. **/
 function main() {
 	for (script in ["Clean", "Build", "Version"]) Sys.command('lix $script');
-	FileSystem.deleteFile("www/index.php");
+	if (FileSystem.exists("www/index.php")) FileSystem.deleteFile("www/index.php");
 
-	minifyFile("bin/php_index.js", Node);
-	minifyFile("www/js/main.js", Browser);
+	minifyFile("bin/php_index.js", true);
+	minifyFile("www/js/main.js", false);
 	Sys.command("npx @cedx/php-minifier --mode=fast lib");
 }
 
-/** Minifies the specified `source` file. **/
-private function minifyFile(source: String, ?destination: String, platform: Platform) Sys.command("npx", [
+/** Minifies the specified source file. **/
+private function minifyFile(source: String, ?destination: String, isNode = false) Sys.command("npx", [
 	"esbuild",
 	"--allow-overwrite",
 	"--legal-comments=none",
 	"--log-level=warning",
 	"--minify",
 	'--outfile=${destination ?? source}',
-	'--platform=$platform',
+	'--platform=${isNode ? "node" : "browser"}',
 	source
 ]);
